crypto_analyzer/requirements.txt
# 核心依赖
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
ccxt>=4.0.0

# 技术分析
TA-Lib>=0.4.28
ta>=0.10.2

# AI模型调用
openai>=1.12.0
anthropic>=0.25.0

# 数据处理
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 调度任务
apscheduler>=3.10.0

# 配置文件
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web框架
flask>=3.0.0
flask-cors>=4.0.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0